wada_brand/requirements.txt
django==4.1.13
Pillow
mysqlclient
django-cors-headers
PyJwt